+++ src/Material.hpp
@@ -3,10 +3,10 @@
 #include <glm/glm.hpp>
 
 struct Material {
-  const glm::vec3 kd;
-  const glm::vec3 ks;
+  const glm::dvec3 kd;
+  const glm::dvec3 ks;
   const float shininess;
 
-  Material(const glm::vec3 &kd, const glm::vec3 &ks, const float &shininess) : 
+  Material(const glm::dvec3 &kd, const glm::dvec3 &ks, const float &shininess) : 
     kd(kd), ks(ks), shininess(shininess) {}
 };
